+++ src/features/auth/index.ts
@@ -1,5 +1,4 @@
 export { AuthPagesLayout } from "./AuthPagesLayout";
 export { Login } from "./Login";
-export { MyTextField } from "./MyTextField";
 export { Register } from "./Register";
 export { RequireAuth } from "./RequireAuth";
