+++ src/features/common/index.ts
@@ -1,2 +1,3 @@
 export { Layout } from "./Layout";
 export { Loader } from "./Loader";
+export { MyTextField } from "./MyTextField";
